api/requirements.txt
psycopg2-binary
flask
logger
flask_sqlalchemy
